requirements.txt
requests==2.26.0
allure-pytest=2.9.44
pytest=6.2.5
